[BugFix] Fix the issue where fixed elements cannot be removed from the page node when multiple ancestor nodes are removed during a single rendering process.
During a single rendering process, when multiple ancestor nodes are removed, the fixed element cannot be removed from the page node. This is because the removal operation of the fixed element is recorded on a node that has been removed, and the action for that removed node does not execute, resulting in the fixed element not being genuinely removed.

Therefore, when we remove an element, we will check if it has recorded any special actions. If so, these actions will be propagated upward.

issue:m-6153270006
AutoSubmit:True
AutoLand:release/3.1