[Optimize][Infra] Create python venv and install dependencies by default.
Summary:
Previously, vpython was used to create the venv and install dependencies such as yaml and requests. This approach not only required the venv creation and dependency installation steps to be repeated every time vpython was invoked, but also faced the issue of vpython not being cross-platform compatible.

Key modifications:
1. Removed vpython and directly created the venv during envsetup, pulling dependencies in one time;
2. The project now directly uses the Python from the venv;
3. Adapted scripts that previously invoked vpython.

AutoSubmit:true
issue: gh-731