[BugFix] Add NDEBUG defines to avoid memory problem.
In some cpp headers, we use NDEBUG define for debuging variables or DCHECK logic, so we should make sure that all pods should use the same NDEBUG define.
For example, if Lynx has NDEBUG define while BaseDevtool not, adoption_required_ in ref_counted_internal.h is not defined in Lynx pod, but it is defined in BaseDevtool.
Therefore the memory size of RefCountedThreadSafeBase will be different, and it may cause memory occupation problem.
It may also cause that DCHECK is available in release version.