[Optimize] Add RunPixelPipeline Api for LynxEngine
in this commit, `RunPixelPipeline` Api is supported for LynxEngine, RunPixelPipeline owns the resolve, layout, flush op process for lynx pipeline.
For now, it's a single api that not be used, we will migrate lynxEngine api to this in later commits.

While After implementing `RunPixelPipeline` Api, lifecycle state should be advanced will pipeline go ahead. And layout is still trigger by element_manager, it should be refactored to be invoked directly by LynxEngine.

issue: f-6577188011
AutoSubmit: true