[Testing][TestBench] Add `forbidTimeFreeze` query for TestBenchReplay
The TestBench by default mocks the JS engine, ensuring that some global objects remain
in a recording state, such as the Date object. The main goal is to recreate the state of
the recording phase as much as possible. However, we need to provide a manual shutdown
capability to correctly verify the behavior of the JS engine during the testing phase.