[BugFix][iOS][Text] Fix the issue that the text still displays after being cleared.
In commit 0c064b02c2193e8303b940d124362be55b0af215, we unified the trigger of text drawing in `onNodeReady`, but this triggered a hidden bug. In the scenario where the text is cleared, when the `onReceiveUIOperation` function is executed , the frame of the UI is non-zero, and a drawing will be triggered at this time, and the previous content will be cleared. After the modification, the drawing will only be triggered in `onNodeReady`. At this time, the frame of the UI is already 0, so it returns directly in advance without clearing the content.

Clear the layer whenever drawing is triggered.

issue:m-6673432609
AutoLand:release/3.3