[BugFix][iOS][Event] Fixed the issue that lynx.resumeExposure on iOS is invalid due to premature calling.
description: This problem reappears in some scenarios on the iOS side where stopexposure is called first and then resumeexposure is called. At this time, when resumeexposure is called, the exposure data source has no data yet, and the stopexposure flag is also true, resulting in the inability to start exposure detection when subsequent exposure data sources are added.

AutoLand: release/3.3

issue: m-6712665030