[BugFix] The creation of PaintingContext should not be affected by the needPaintingContextProxy check
In Embedded mode, entering the Engine reuse process will skip the creation of PaintingContext. However, the creation of PaintingContext here should not be checked by the needPaintingContextProxy flag. Otherwise, PaintingContext may not be created correctly in some cases. After the fix, it remains consistent with the historical logic.

issue:m-6739882771
AutoLand: release/3.4