[Optimize] Add `SetValueUniqueKey` for `lepus::Dictionary`
When an empty `lepus::Dictionary` is about to have a set of data added to it, and it is guaranteed that there will be no key conflicts, the `insert_unique` mechanism can be used as an optimization. The `insert_unique` method does not perform a key lookup before inserting the data. This is often used to copy data from the original map to the newly created `lepus::Dictionary`, because the original map is impossible to have conflicting keys, thus allowing the `insert_unique` optimization. Currently, `std::unordered_map` does not support this capability.

This optimization will take effect after the refactor is completed.