[Optimize] Introduce `BundledOptionals` class to base library
`BundledOptionals` is a template type that dynamically manages multiple optional fields. These fields are dynamically allocated on the same block of memory and moved as fields are added or deleted. Its template parameters and interface are similar to `std::variant`, but `std::variant` only allows one type to be stored at a time.